refactor: move and use from-address utils to shared folder
